added catalog model and repo with updates to the id type for all models

+++ src/main/java/com/jonescorp/shoppy/model/Transaction.java
@@ -22,10 +22,10 @@ public class Transaction {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer transactionId;
+    private Long transactionId;
     private Date date;
     private ArrayList items;
-    private Integer userId;
+    private Long userId;
     private String userName;
 
 }
